Save Task eingebaut. Nach Stefan. Speichert die Classifier beim Beenden. Ausprobieren mehrerer Parameter. Einfachste Version jetzt. Nachbau nearest Pill Pacman mit xcs. Fehler beseitigt.

+++ PacMan/src/game/entries/pacman/group4/IEnvironmentObserver.java
@@ -24,6 +24,6 @@ public interface IEnvironmentObserver {
 	 * @param time vergangene Zeit
 	 * @return Integer gibt den Reward an.
 	 */
-	public int getReward(Game game, long time);
+	public double getReward(Game game, long time);
 
 }
